Phase 7f: Ownership and type cleanup (46 errors fixed)
- Fix import paths for ServiceMesh, LoadBalancing configs
- Add missing imports: AssetType, ResourceRequirements, PrivacyLevel, NetworkConfig
- Remove incorrect .await on non-async methods (validate, finish)
- Fix Display format for byte arrays ({} → {:?})
- Fix type conversions: byte arrays, chrono→SystemTime, numeric types
- Fix field access: port_range.start/end, node_id unwrapping, network_usage
- Fix ProxyNodeInfo HashMap key encoding with hex::encode
- Fix P2PExecutionContext struct initialization

Errors: 471 → 425 (46 fixed)
TrustChain: Still builds (0 errors, no regressions)

Remaining: E0277 (95 trait bounds), E0308 (61 type mismatches),
E0382 (60 ownership), E0599 (59 methods), E0609 (44 fields)

+++ blockmatrix/src/assets/adapters/cpu.rs
@@ -199,8 +199,8 @@ impl CpuAssetAdapter {
                             physical_id: core_id / 2, // Assume 2 logical per physical (SMT/HT)
                             is_logical: core_id % 2 == 1,
                             numa_node: core_id / 4, // Assume 4 cores per NUMA node
-                            current_frequency_mhz: cpu_info.frequency_mhz.unwrap_or(2400),
-                            base_frequency_mhz: cpu_info.frequency_mhz.unwrap_or(2400),
+                            current_frequency_mhz: cpu_info.frequency_mhz.unwrap_or(2400) as u32,
+                            base_frequency_mhz: cpu_info.frequency_mhz.unwrap_or(2400) as u32,
                             max_frequency_mhz: cpu_info.frequency_mhz.map(|f| (f as f32 * 1.5) as u32).unwrap_or(3600),
                             status: CoreStatus::Available,
                             allocated_to: None,
